Refactoring of eye, ones, and zeros
