+++ ChangeLog
@@ -1,3 +1,13 @@
+2016-10-24  Nathan Russell  <russell.nr2012@gmail.com>
+
+	* inst/include/Rcpp/vector/MatrixBase.h: Change sugar 
+	functions eye(), ones(), and zeros() into static methods 
+	in MatrixBase 
+	* inst/include/Rcpp/traits/one_type.h: Idem 
+	* inst/include/Rcpp/traits/traits.h: Idem 
+	* inst/unitTests/cpp/Matrix.cpp: Idem
+	* inst/unitTests/runit.Matrix.R: Idem 
+
 2016-10-23  Nathan Russell  <russell.nr2012@gmail.com>
 
         * inst/include/Rcpp/sugar/matrix/eye.h: New functions
@@ -7,11 +17,11 @@
         * inst/unitTests/runit.sugar.R: Idem
         * inst/unitTests/runit.dispatch.R (test.ExpressionVector): Use
         expression rather than parse, correct typo
+
 2016-10-22  Qiang Kou  <qkou@umail.iu.edu>
 
         * inst/include/Rcpp/hash/IndexHash.h: change hashing function to return unsigned int
         * inst/include/Rcpp/hash/SelfHash.h: Ditto
-
 
 2016-10-21  Qiang Kou  <qkou@umail.iu.edu>
