+++ inst/include/Rcpp/sugar/matrix/matrix_functions.h
@@ -29,6 +29,5 @@
 #include <Rcpp/sugar/matrix/upper_tri.h>
 #include <Rcpp/sugar/matrix/diag.h>
 #include <Rcpp/sugar/matrix/as_vector.h>
-#include <Rcpp/sugar/matrix/eye.h>
 
 #endif
