+++ inst/unitTests/runit.sugar.R
@@ -1610,95 +1610,4 @@ if (.runThisTest) {
   
     }
     
-    
-    ## 23 October 2016
-    ## sugar eye function
-    test.sugar.eye <- function() {
-        
-        checkEquals(
-            dbl_eye(3),
-            diag(1.0, 3, 3),
-            "eye - numeric"
-        )
-        
-        checkEquals(
-            int_eye(3),
-            diag(1L, 3, 3),
-            "eye - integer"
-        )
-        
-        checkEquals(
-            cx_eye(3),
-            diag(1.0 + 0i, 3, 3),
-            "eye - complex"
-        )
-        
-        ## diag(TRUE, 3, 3) was registering as 
-        ## a numeric matrix on Travis for some reason
-        mat <- matrix(FALSE, 3, 3)
-        diag(mat) <- TRUE
-        checkEquals(
-            lgl_eye(3),
-            mat,
-            "eye - logical"
-        )
-    }
-    
-    ## sugar ones function
-    test.sugar.ones <- function() {
-        
-        checkEquals(
-            dbl_ones(3),
-            matrix(1.0, 3, 3),
-            "ones - numeric"
-        )
-        
-        checkEquals(
-            int_ones(3),
-            matrix(1L, 3, 3),
-            "ones - integer"
-        )
-        
-        checkEquals(
-            cx_ones(3),
-            matrix(1.0 + 0i, 3, 3),
-            "ones - complex"
-        )
-        
-        checkEquals(
-            lgl_ones(3),
-            matrix(TRUE, 3, 3),
-            "ones - logical"
-        )
-    }
-    
-    ## sugar ones function
-    test.sugar.zeros <- function() {
-        
-        checkEquals(
-            dbl_zeros(3),
-            matrix(0.0, 3, 3),
-            "zeros - numeric"
-        )
-        
-        checkEquals(
-            int_zeros(3),
-            matrix(0L, 3, 3),
-            "zeros - integer"
-        )
-        
-        checkEquals(
-            cx_zeros(3),
-            matrix(0.0 + 0i, 3, 3),
-            "zeros - complex"
-        )
-        
-        checkEquals(
-            lgl_zeros(3),
-            matrix(FALSE, 3, 3),
-            "zeros - logical"
-        )
-    }
-    
 }
-
